requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
channels==1.1.6
constantly==15.1.0
daphne==1.4.2
Django==2.1.7
djangorestframework==3.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack-python==0.5.6
PyHamcrest==1.9.0
pytz==2018.9
redis==2.10.6
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
